Doctest fixes and Jorgen's unicode patch for %timeit.
The doctest changes make some examples not be loaded as doctests,
because for now they break the machinery.  Later we'll refine our
infrastructure, but for now I just want to disable them as tests while
leaving their contents in place.

Jorgen's patch fixes a unicode issue in %timeit which was showing up
under Win32.