Updating the Sphinx docs in preparation for the release.
I am going through the docs to make sure that i) there are no problems
with the generation of the the Sphinx docs and ii) that our docs
are an accurate reflection of the current state of the code.

As of this commit, I have made a first pass at everything *except*:

* Everything in the docs/source/interactive dir.  This corresponds
  to most of the old IPython docs about the ipython shell.
* The sections on the multiengine and task interfaces.

Of course, this is just a first pass and we likely have lots more
to do, but at least, our documentation exists and is no longer
misleading or wrong.