Moved the traitlets based config to sandbox (created it).
Stefan's prototype new config system that is based on ideas from
traits, depended on Python 2.5.  We have not done a thorough code
review of this code and I don't think we are even using it yet.

I have moved all the code from IPython.config to the top-level
sandbox.  I have begun to remove all references to these modules
(just a few imports), but I still need to do lots of testing
to make sure there are no lurking problems.