Update copyright/author statements.
- Updated copyright statements to use the new form:

#       Copyright (C) 2008-2009 The IPython Development Team

I left the old notices in place (just updating the year in some cases),
because as far as I know, old copyright statements are not meant to be
retroactively modified.

- Also, on most files, replaced __author__ marks with an 'Authors' section
in the module docstring.  This reduces top-level code in the module, while
ensuring that the Author(s) get properly acknowledged in auto-generated API
docs (sphinx doesn't read __author__ marks, but it will include the module
docstring).

I only left a few in place for very old files that we ship externally, and
for those by Laurent: he had his authorship mark both in the docstring and
in __author__ variables, so I think out of courtesy it would be better to
ask him about it on the list.  All the others were I found regular
__author__ variables, I moved them to the main docstring.