Fix problems with side effects and python 2.6.
The isolation decorator wasn't working correctly (side effects were leaking
into other tests by breaking the user namespace).  Rewrote it to instead use a
simpler but more robust strategy (that is also faster).  See detailed comment
in source for more.

I discovered that some tests were failing when running with python 2.6 because
of changes in the behavior of int and the new print-as-a-function that Wx seems
to activate.  Updated the code to cope with both 2.5 and 2.6.