Cleaned up release tools directory.
Converted almost all to python scripts and made toollib to collect common
functions and avoid repetition.

Properly commented and documented what each script does.

The run_ipy_in_profiler one seems broken, I'm not sure what to do with it.
We need to either fix it or remove it later, but it's not critical for 0.10.