Protect against absent global IPython instance.
This basically applies most of the changes to ultraTB made by Brian in the
reorg branch here:

http://bazaar.launchpad.net/~ipython-dev/ipython/module-reorg/revision/1225

To our duplicate copies of ultraTB (the duplication is already gone in the
reorg branch).

This now means that the trial-based tests run fine if executed as

trial IPython

OR 

trial IPython.kernel

where as the latter form was failing before.  The reason why the former used
to work was that a hidden global ipython is created, *upon import*, by
ipdoctest.  That will stay for now, but I've made it bug

https://bugs.launchpad.net/ipython/+bug/409096

marked as critical for 0.11 that I'll need to fix.