Changed how config files are loaded.
Previously, we were loading a config file by exec'ing the file in 
an empty Config instance.  That was a bit dangerous, so now there
is a get_config function that should be called to get the Config
instance.  Then attributes can simply be set on it.  

The config files have been updated to reflect this change.