Fixed bug in ultratb.py and io.py.
Previously :class:`IPython.utils.io.Term` had a :meth:`Term.writeln`
method that :mod:`ultratb.py` was using. But, when a real stream
or files is used in :class:`IPython.utils.io.Term` it doesn't have this 
method, so we were getting an error when the IPython Sphinx directive
was run. I removed this method and made :mod:`ultratb.py` use
just ``write``.