Add pexpect version 2.3 (revision 507) to IPython.external.
We will only use our copy if the system-installed one isn't found.

This lets us run the core of ipython on top of the stdlib, while
having far better subprocess control than we otherwise would on posix.
On windows, pexpect doesn't exist, so the subprocess situation is
still not ideal.

pexpect is MIT-licensed.

For more information on pexpect:
http://www.noah.org/wiki/Pexpect