Make our temp file mixin unittest-friendly.
This does not break any nose compatibility, but it lets us use our
mixin with normal unittest.TestCase objects in addition to
nose-oriented tests.  We're trying to lower our dependency on
specifics of the nose api, so that in the future we might be able to
run without nose at all (as unittest improves).