Simplify completer handling by isolating readline-specific logic more.
Also, I removed a hack we were using of storing the completer globally
inside sys.  This was used for the rare case of embedding
ipython-inside-ipython, something which is more of a party trick.  If
we ever want to support that well, we can write a cleaner solution
than polluting sys with 'ipcompleter'.  But for now, I'm trying to
simplify our completer machinery and remove buggy and/or error-prone
hacks.

Note that the normal use of embedded ipython inside of user's programs
isn't affected by this.