Speedup builtin_trap enter/exit by reducing object creation.
Most of the work that was being done on *every* enter/exit cycle could
be done statically only once, at object creation time.  This trap is
used on every single code execution, so we need it to be fast.  In the
long run we probably want to rethink this system altogether, but for
now at least we want it fast.