Limit special-casing of _ variable to doctests.
In doctests, _ *must* be special-cased and removed from the user's
namespace because otherwise Python won't set it.  But we were doing
this special-casing unconditionally, making it impossible to write
unittests that checked the _ variable after cell execution.

This commit makes the special-casing of _ apply only in the doctest
execution (where it's needed) and leaves normal unittests alone.