fix various tests on Windows
Windows had a variety of mostly path-related test failures remaining.

Most of these were Windows-excluding assumptions in the tests themselves.  The one change to the code itself is stripping quotation marks on Windows in magic_cd.  os.chdir doesn't like the quotes on Windows.

absolute-path completions are now listed as a KnownFailure on Windows