allow engines to wait for url_files to arrive
if connection info (url, ip, etc.) is passed explicitly, the engine will not wait for the file.

If the file never arrives and the connection info was not specified manually, the app will abort.

This is primarily useful in batch-systems where engines might start before the controller has a chance to write connection files. (delay does not help in these cases, because it could be hours before jobs start)