Update docs about autocall and remove manual copy of cmd-line flags.
A manual copy of all the flags kept in the docs will always grow
stale, so we might as well simply indicate to users how to get the
listing at runtime.  If there's real demand for having this in the
docs, then it should be auto-generated so at least it's up to date
with the code.