Make all the static files path absolute.
This is necessary because, if we don't do it all the paths will be
relative to the current URL and whenever we want to render a notebook
view from the paths other than the root URL of /, the static files path
get appended to that path from where we are asking for static files and
since static files are not mapped to those URLs we get HTTP 404 on all
those static files.

For example: If we render a notebook view from say /kernels/kernel-4444.json,
the static files will be looked at /kernel/static/... without this patch.
But after applying this, irrespective of which URL we are at, the static
files will be looked up at /static/...