Fix remaining test failures.
This completes the decoupling of the Magic class from the main
InteractiveShell one.  Finally! (it's been only 10 1/2 years...)

At this point, all tests pass (modulo two failures in zmq that were
there before and are not related to the magics work).

We can now start the rest of the work for cell level magics, including
a cleaner magic management architecture and breaking them up into
several classes.