Settle on cleaner API for magic registration.
The official API will be:

- ip.register_magics(*args): for registering one or more classes or
instances that subclass the main magic.Magics class.  This will be the
*only* method for registering magics that have a signature f(self,
line,...).

- ip.function_as_magic: for registering one-off magics made from a
standalone function with the signatures f(line), f(line, cell) or
f(line, cell=None).

We will support, for backwards compatibility, the old ip.define_magic,
but it will print a deprecation warning.