Fix bugs in completer.py with incompletely-implemented property.
From a docstring that ended mid-sentence, it's pretty clear that we'd
forgotten to finish the implementation of the completion splitter with
real properties.  I have no clue how the tests were passing, as there
was also a nasty bit of state mangling being done by one of the tests;
finished the implementation and fixed the tests.