'Confirm kernel restart' updated on tab change
The way I originally wrote the code, it was possible for the
"Kernel" -> "Confirm kernel restart" checkbox to get out of sync with
the currently active tab. It turns out that QTabWidget has a
'currentChanged' slot that can be leveraged to update the checkbox to
properly reflect the currently active tab widget.