Line continuations now terminate after one blank line (#2108)
Previously, we had

In [1]: 1\
...:
...:
...:
...:

In other words, no amount of blank lines would terminate after a line
continuation, in contrast with regular Python:

>>> 1\
...
1

This made things really annoying when I typed \ instead of a newline--quite
easy to do since they are right next to each other on the keyboard.

Now, we have

In [1]: 1\
...:
Out[1]: 1

This also fixes another related behavioral difference between IPython.  If a
space follows a line continuation character, it should be a
SyntaxError("unexpected character after line continuation character"), even if
the line is otherwise continuable, according to regular Python (e.g., `1 \ `
or `(1 + \ `).  This now consistent between the two.

Closes #2108