Add true HTML export
This adds full HTML export with notebook css, etc.

todo:

Figures are inlined, and foo_files/image.png mode should probably be available as an option.

MathJax is still loaded from a CDN, and should probably be configurable (or disabled).

We could probably trim most of the CSS, so we just have what we need, but that would mean maintaining some synchronization with upstream CSS probably.