Backport PR #2214: use KernelApp.exec_lines/files in IPEngineApp
This change means that all your startup/exec_lines config is inherited by engines.

should be backported for 0.13.1, because engine startup code is impossible to configure in 0.13.

The real long-term fix for this is to make IPEngineApp a subclass of IPKernelApp, but this should be an adequate band-aid until then.

closes #2213