iostream._newbuffer starts auto-flush clock
This means that the timer runs between flush calls, rather than the previous logic where the timer started on the first print after each flush.  The problem with the previous logic was that it meant the first message after any flush could *never* be displayed immediately without an explicit manual `stdout.flush()`.

The disadvantage of the new approach is that most cells with two or more print statements will result in at least two stdout messages, because when used interactively the first print statement will ~always trigger the flush_timeout logic.