Backport PR #2384: Adapt inline backend to changes in matplotlib
Matplotlib recently merged https://github.com/matplotlib/matplotlib/pull/1125 that makes it simpler to use
objective oriented figure creation by automatically creating the right canvas for the backend. To solve that
all backends must provide a backend_xxx.FigureCanvas. This is obviosly missing from the inline backend.

The change is needed to make the inline backend work with mpl's 1.2.x branch which is due to released soon. Simply setting the default canvas equal to a Agg canvas appears to work for both svg and png figures but I'm not sure weather that  is the right approach. Should the canvas depend on the figure format and provide a svg canvas for a svg figure? (Note that before this change to matplotlib the canvas from a plt.figure call seams to be a agg type in all cases)

Edit: I made the pull request against 0.13.1 since it would be good to have this in the stable branch for when mpl is released.
Just let me know and I can rebase it against master