Remove self._anchor
Bradley Froehle recommended removing self._anchor and just
requesting the anchor text for a position as it was needed.

He also pointed out that Qt handles hiding the tooltip for
the empty string and suppressing redraw messages which
simplified the event handler code.

The only problem I found is when updating my test code the tooltip
takes a moment to disappear and caches the text value for a time
longer than a function call. So I can't verify that the tooltip is
cleared when the mouse moves off the anchor.