Remove Equation References/Numbering, Fix Bugs
Equation References and Numbering are not going to be trivial to add,
so the code has been removed for now.  Important fixes include
no-MathJax support (previously, the code was failing), and the
generation of unique ids for the rendering content in each cell,
tremendously speeding up MathJax rendering.

I am still not rendering placeholder text.