this should address the failure in #2732
i am still very concerned that I can't test this on windows - I'm not used to developing for windows, so not sure that these links will work even if the test passes on windows.

is there a good AWS AMI to use for windows testing? i can run this in my own AWS account.