truncate potentially long CompositeErrors
If you have a typo in a task on 100 engines, the local exception will be a bit ridiculous.

This defaults the truncation to 4 tracebacks (could be even fewer, or even one?),
and you can set the tb_limit attribute of CompositeError to adjust this.

Prompted by [this SO question](http://stackoverflow.com/questions/14635935/silence-or-condence-ipython-parallel-exceptions).