Respect DB_IP and DB_PORT in mongodb tests.
It is usually preferred to run tests on a test-local database with
dedicated mongod instance rather than using the system-wide database.
In order to support that, an easy way of providing the proper port
is handy.

For example, pymongo test suite uses DB_IP and DB_PORT environment
variables to allow providing an alternate mongod instance. Using
the same variable names in IPython's test suite seems a good idea.