Backport PR #2662: qtconsole: wrap argument list in tooltip to match width of text body
previously, a function with a long argument list would produce
a very wide tooltip, hurting readability.

Since the width of the docstring body is chosen by the developer
whereas the formatting of the argument list is currently not, it's
reasonable to wrap the argument list width to the maximum of
a) 80 characters, or b) maximum length of a line in the docstring body.

This improves readability, without unduly affecting the appearence
of the docstring body itself.

closes #2661

I'm happy to add any additional tests or make any changes required to get this merged.