fix test_find_cmd_python
It's a bit unclear what should be done here.
'python' is short-circuited in find_cmd to sys.executable.
This commit does the same short-circuit with basename(sys.executable), which can be any of Python, python, python3, python3.exe, etc.

An alternate fix would be to change the test back to testing what the code actually did.