remove pdf from display data_priority
rational is to have 1 to 1 mapping between stored data in notebook
and branches in templates.

If the fie need to be converted in another format then the
configurability of templates for filename in the notebook should allow
to do it.

basically if I have a foo.svg figure I want to include in the notebook,
the {{ svg block }} shoudl be able to include a name that is

foo.svg, or foo.png or foo.tiff or foo.gif