Remove 5s wait on inactivity on GUI inputhook loops
The 5s (and 1s) waits were originally added in commit 5074878, but the 5 second wait meant if you left the console for 5+ minutes idle, it would
take up to 5 seconds for a response to a keypress. This tradeoff of CPU
cycles for battery life seems too far.

Note that commit 5074878 was originally for wx, glut and pyglet are based
on the wx version and came into existence after commit 5074878.