Don't check for wx in the test suite.
An interaction between wx and threads was blocking certain Unix signals,
causing an unrelated test to fail for me.

We were only checking for wx to exclude the inputhookwx module from the
test run. It contains no tests anyway, so we may as well just exclude it
in all cases.