Backport PR #4183: ESC should be handled by CM if tooltip is not on
This PR fixes an issue on a key-press event with ESC  in code cells, allowing the event to propagate to CM if tooltip is not on.

Currently, key-pressing ESC in a code cell simply removes tooltip, and the event is ignored by CM.   I can not get out of insert mode when I am using Vim key binding.