Backport PR #4232: no highlight if no language specified
This matches behavior of places like GitHub, where if you do not specify a language, e.g.

    ```
    no_highlighting("here")
    ```

There is no syntax highlighting. You must do

    ```python
    def foo(a="hi"):
        return 5
    ```

to enable highlighting in a particular language.

With configuration as it is right now in IPython, it is impossible to disable syntax highlighting in markdown blocks.