Backport PR #5459: Fix interact animation page jump FF
Firefox doesn't render images immediately as the data is available.  When animating the way that we animate, this causes the output area to collapse quickly before returning to its original size.  When the output area collapses, FireFox scrolls upwards in attempt to compensate for the lost vertical content (so it looks like you are on the same spot in the page, with respect to the contents below the image's prior location).  The solution is to resize the image output after the `img onload` event has fired.

This PR:
- Releases the `clear_output` height lock after the image has been loaded (instead of immediately or using a timeout).
- Removes a `setTimeout` call in the `append_output` method.
- `clear_output` in zmqshell no longer sends `\r` to the stream outputs.

closes #5128