Backport PR #5488: Added missing require and jquery from cdn.
For some reason (I suppose some changes at the css level) the font size inside the input cells was fixed at 14 px... making the fonts really small in the reveal slideshows. This is really annoying...
As a plus, I have also added the missing calls for require and jquery (as the full html template does).
I think these fixes belong to 2.0, but I also know we are on the edge... so I hope to get it inside :wink:

Cheers.