Backport PR #5534: cleanup some `pre` css inheritance
We had various overrides of the bootstrap defaults for font-size and line-height every time we made a pre tag, and then some overrides of our overrides when we wanted something slightly different.

This moves the most basic overrides to apply to all pre tags, reducing the number of downstream changes we need.

Should hopefully replace the CSS changes in #5488