Do not create checkpoint_dir relative to current dir
FileNotebookManager.create_checkpoint attempted to create checkpoint_dir
relative to the current working directory, not the notebooks directory set
in NotebookManager.notebook_dir. FileNotebookManager.get_checkpoint_path
already creates checkpoint_dir.

Fixed already in 19651d2, but this can be used as a small bugfix for 2.x,
in case that is not being backported.