Backport PR #6005: Changed right arrow key movement function to mirror left arrow key
Seems to solve Issue #5926 on this machine, and passing the test file locally.  Changed from `cursor.movePosition` to `self._control.moveCursor`, the latter is what the left-arrow key uses.  Also removed line 1373 which seems unnecessary and which prevents the cursor from moving at all.

I'm not certain how to further test this to make sure nothing was broken.