Tree Icon Inversion
Tiny change, inverts the colors of the tab icons in tree.html to make
them easy to spot relative to notebook tabs.  Added to my own install to
keep track of many tabs.

Not sure if everybody would like this.  I'd welcome suggestions for the
best way to make the change optional.