allow splitting and merging of heading cells
I consider it a bug that you couldn't merge with heading cells, and that
you couldn't split them, either. (So much so that I thought it was a bug
in ipython-vimception when I ran into it).

This change removes that limitation, so heading cells are on par with
the other cells in terms of the kinds of manipulations one can carry out
with them.