Fixed whitespace bug when pasting IPython continuation prompt.
Fixes a SyntaxError that arose when pasting code containing the IPython
continuation prompt without a trailing white space, such as:

In[1]: if True:
   ...:    print "No space after ':' on next line."
   ...:

This might arise if the user only copies up to the end of the last
prompt, without copying the trailing whitespace.

This commit adds a regression test and fixes the regular expression that
strips continuation prompts, making the whitespace optional.