Fix incompatibility with the python-future package
The python-future ( http://python-future.org/ )  package provides a dummy version the _thread module for python2. This causes the dummy library to be loaded instead. The effect is that for example the 'Interrupt' button in ipython notebook doesn't work.

This fix ensures that we try to import the python 2 code first, thus preventing the _thread module from being imported when the future package is installed.