copy timeit.Timer.timeit from CPython 3.4
we were relying on self.inner for %timeit,
which is an undocumented implementation detail
of Python, and no longer shared by PyPy.

This copies the CPython implementation,
so it is no longer an undocumented attribute.

It will likely not be optimal for some PyPy timings,
but at least it works.