Reverse hscrollbar min-height hack on OS X
OS X has optional behavior to only draw scrollbars during scroll,
which causes problems for CodeMirror's scrollbars.
CodeMirror's solution is to set a minimum size for their
scrollbars, which is always present.

The trade is that the container overlays most of the last line,
swallowing click events when there is scrolling to do,
even when no scrollbar is visible.

This reverses the trade, recovering the click events at the expense of
never showing the horizontal scrollbar on OS X when this option is enabled.