imp module deprecated in 3.4
We still use imp.find_module,(in IPython/utils/module_paths.py)  but
this as no newer drop-in replacement as far as I understand.

Some alternative suggested here:

https://docs.python.org/3/library/importlib.html#importlib.abc.Finder

Deprecated since version 3.3: Use MetaPathFinder or PathEntryFinder
instead.

The various error messages are not clear of the exact version where
this is deprecated.