added supressExceptions paramter to FullEvalFormatter
In order to maintain the current behaviour of FullEvalFormatter (it raises exceptions if it's `eval` fails), a `supressExceptions` parameter is added, defaulting to False. In the DollarFormatter, used for cell magic parsing, it defaults to True. This allows some expressions to be evaluated when there is an exception in another expression.  See #8492 for a discussion...

It only catches NameErrora - perhaps it'd be better to catch all exceptions? 

It may also be a good idea for cell magics to emit a warning if an exception is raised during variable substitution. For that, we'd have to collect the exceptions here and somehow pass them on to the cell magics...