Remove old deprecated extensions.
All theses extensions are already in their own repository.

Local Rmagic could in any cases not be imported as %load_ext would raise
if rpy2 not available.

Cython and sympy where still trying to do the right things, which I
removed, and converted to warnings. (cannot use DeprecationWarning as
they would be hidden) and no way to get the right stacklevel as this is
magic that load the ext.