Merge in qtconsole/qt.py API-loading logic.
IPython.external.qt was moved to qtconsole.qt, but was still imported
resulting in ImportErrors when trying to use Qt. This commit brings
the logic back into this file. Deferring to qt.py was a bit silly because
both qt.py and qt_for_kernel.py do basically the same thing but in
slightly different ways. Moreover, the assumption of qt.py was that if
the QT_API environment variable is set, we must be in an ETS environment
and need to comply by the ETS import logic. This is not valid, as
ipython sets the QT_API environment variable to communicate intent as to
Qt version, in IPython/lib/inputhook.py:385
@inputhook_manager.register('qt5')
class Qt5InputHook(Qt4InputHook):
    def enable(self, app=None):
        os.environ['QT_API'] = 'pyqt5'
        return Qt4InputHook.enable(self, app)

Now the logic is centralized.