Move most readline code from InteractiveShell to terminal subclass
This is the first stage of my effort to use prompt_toolkit for the
terminal interface.

My plan is to create a new subclass of InteractiveShell which deals with
the terminal using prompt_toolkit instead of readline. To facilitate
this, I am first trying to separate as much of the readline code as
possible out of InteractiveShell to the TerminalInteractiveShell
subclass (I don't intend the new subclass to inherit from TIS).

I haven't moved any of the config options relating to readline yet,
because moving them will break some people's existing config files. But
if we're switching to prompt_toolkit by default, readline related config
is going to be broken anyway.