Explicits deprecation warnings removal versions.
Otherwise we never know when a feature will be removed.

Nothing prevent us to keep longer, but at least giving a version number
instead of "Next Major release" give a hint to users. And also when we
want to cleanup later, we don't have to wonder when the deprecation was
added.