Deactivate prompt transformers if the first line matches certain patterns
Deactivate the >>> prompt stripper if the first line starts with % or !,
because those are obviously IPython syntax.

Disable the IPython 'In [1]:' prompt stripper if the first line starts
with %%, because we don't want to interfere with the copy of a cell
magic.

This is all 'what did the user mean' guesswork, but I think this makes
our guesses slightly smarter.

We actually had one test that checked that this wasn't happening; I've
adjusted it.

Closes gh-8791