Allow Embeded instances to raise on exit.
This is use full in case where an IPython embed instance is created in a loop:

    from IPython import embed
    while True:
        embed(header="You will be stuck there")

Now using `%raise_on_exist True` in an embeded shell allows the embeded
instannce to raise on exit.

Potential improvement would be flags to pass a custom exception type as
well as a custom message to raise.

Closes #9149