remove setuptools workaround for platform_python_implementation
This was a workaround for old setuptools, but we've added a dependency on 18.5,
so it shouldn't be required anymore.

Setuptools 20.2 has removed python_implementation (perhaps unintentionally),
so the workaround actually *breaks* on more recent setuptools.

We should have only made this patch when setuptools needed it, but we can remove it outright, now.