"smart command mode" for ipdb (like pdb++).
If e.g. "c" is a variable currently visible, then "c" should print it
instead of running the "continue" Pdb command.  This idea was first
implemented by pdb++.

It is possible to force the standard pdb behavior (bypass variable
checking) by prefixing the command with "!!", e.g. "!!c" to continue.