Properly complete arguments to %cd.
Rely on `arg_split` to re-split the entire line (up to the cursor) so
that directory names that include readline separators are properly
handled.

Because the completed symbol may not be the one that
`IPCompleter.complete` (which relies on naive splitting) believes to be,
it is necessary to tag each candidate completion with the symbol *it*
completes.  (Note that this symbol is not even necessarily a prefix of
the "original" symbol, e.g., on Windows, paths may get a double-quote
prepended upon completion.)  This is done using an attribute on an str
subclass.

Multiple completions may end up completing symbols of different length,
so the `_pad_completions` function is used to "pad" the completions,
pretending that they are all completing the same symbol.

Because completions no longer necessarily start with the completed
symbol, it is up to the individual completers to filter out
non-matching completions.