Release 5.0.0b4
What's new since b3:

Small delay since last beta release, sorry about that, a few reasons though:

- We / I slowed down a bit on the beta release pace, there was a few
  bugs/issues/User interface annoyance that needed a new version of
  prompt_toolkit to be released, there was no reasons to make extra-beta
  without this new version released.

- We backed-down on some changes on which we had a few disagrements (more below).

- I got sick, and had to catch up with backlog.

Anyway, Jonathan release prompt_toolkit 1.0.1 and 1.0.2, so even if you
don't try this new beta, upgrading prompt_toolkit will fix some of your
issues.

== Move back `TerminalInteractiveShell` to it's old place.

This is a bigger breaking change since 5.0b3, the
`TerminalInteractiveshell` moved back from `IPython.terminal.ptshell` to
`IPython.terminal.interactiveshell`, if you've updated your project
recently to adapt to this change we're sorry, but despite the fact that
the version pre 5.0 and post 5.0 are relatively different the cost of
conditional import for project depending on us appeared to be too high.
So it's now easier to migrate from 4.0 to 5.0 as the class have the same
name, and same location

== Option name and default changed.

`TerminalInteractiveShell.display_completions_in_column` is now gone. It
was not present on 4.x so no API breakage there, and is now replaced by
`TerminalInteractiveShell.display_completions` and is a enum that gained
a 3rd mode for the completer: `readlinelike` for those of you that
regret readline. This give us more flexibility for further options.
Would appreciate testing of this new layout from vi user. The two other
mode now being `column` and `multicolumn`.

By popular request, `multicolumn` is not the default value for the
previous option.

== bug fixed:

- quit/exit broken in ipdb
- Copy/Past broken on windowm
- Unicode broken on windows
- function signature garbled when using `object?`
- issue with paging text with `?`
- completer could get stuck.

See the complete git log for more informations.