Remove the "autoedit_syntax" feature.
IPython used to (a long time ago, in a galaxy far far away) have the
ability to automatically open an editor in case a wild syntax error
appears. The configuration option to enable that was not working for a
few years, and apparently we by mistake re enabled it, to discover that
the feature is actually broken.

So this plainly remove the code to support this feature, at the
exception of the `fix_error_editor` hook. Indeed it is public API, so
for now  as it seem to be used only for this feature, we'll just raise a
UserWarning (there is roughly 0 chance of this being tested as it's used
mostly interactively, so DeprecationWarnings would be unseen).

We'll remove later if no complaints

Closes #9603