Restore the ability to use %colors to switch terminal theme.
And restore previous coloring of prompts on classical color themes.
Unlike 4.x this will **not** change with your terminal background.

If the hightligh_style of TerminalInteractiveSHell is set to `legacy`
then it uses the value of TermianlInteractiveShell.colors to select the
theme:

monokai for darkbg/linux (by decision of BDFL), and old prompt values.
default for lightbg

Closes #9648