Backport PR #9660: Don't rely upon traitlets copying self.config
when saving CLI config to reload later for highest priority.

traitlets 4.1-4.2.1 don't recreate self.config in update_config (rightly
so), so when saving CLI config for later re-loading, make sure it's a
copy so it doesn't get changed during config file loading, which defeats
the purpose.

To me, this is the real fix for ipython/traitlets#248