Fix broken links on install/index.rst
Currently the links in the section summary on the index.rst/index.html
file are broken. Can be reproduced by going to
http://ipython.readthedocs.io/en/stable/install/index.html and clicking
on *installing IPython itself*, or *kernels for Jupyter* links.

Specific changes
----------------
* Added a reference label to the `install.rst` file
* Modified links from external link format to Sphinx arbitrary location
cross-referencing format in the `index.rst` file

Testing
-------
Changes have been tested with a local sphinx build through the supplied
makefile and specific links touched are fixed. Tested a few possibly
overlapping links (Jupyter:install for instance) and they seem to be
unaffected by the change.