Use 'Linux' (dark bg) colours for Windows
This puts the colours in inspect output and tracebacks back to what
they were in IPython 4.x, but leaves the prompt colouring as in 5.0.
I tried changing the colour scheme to 'Linux' entirely for Windows,
but that selects Monokai as the theme for prompt_toolkit, which looks
pretty horrible in 16 colours (at least to my eyes).

This is admittedly a hack, but hopefully our legacy colour system is
on the way out anyway.

Closes gh-9723