Switch master to development of 6.0
I'm thinking about this because we've just released 5.1, but I'm happy
for it to wait a bit longer if we prefer. As planned, 6.0 will require
Python 3, while we will continue to support 5.x for some time for Python
2 users.

Before merging this, we should send a message to the mailing list for
any Python 2 users following master, and make a 5.x branch off master.