Set all SIP APIs
After QT is imported the SIP API functions can no longer be
called to set the QT API. This change ensures that all of
the APIs are either set to 1 or 2 so that other packages
that may be imported after this function have a consistent
interface. This is consistent with other packages such as
pyside and is particularly relevant for IPython as the
%matplotlib magic means that one of the first set of imports
made in an IPython session is often from the %matplotlib
magic.