Backport PR #10150: Always run the OS X loop even if no windows.
Otherwise this can trigger infinite Python-Icon-In-Dock bouncing.
See  10137. I will guess that this is because application on OS X may
not have windows and still need to process events.

It may be that an alternative is to run the loop only once the first
time, but I'm unsure.

----

Ping  minrk : I have no clue what I am doing.