First step in reintegrating Jedi
If Jedi is installed expose a private API use it with prompt toolkit.
Jedi does not _yet_ provide all the completion IPython has, so this is
still a bit awkward. In order to debug this (and see what is Jedi
provided we for now inject a fake Jedi/IPython delimiter in the menu.

Jedi completion and this behavior are enabled by default, but could
likely be opt-in.

Add also a number of debug flags to be able to track why jedi is not
working, and/or what completions are found by IPython and not Jedi.
That should give us a bit of heads up and feedback to know whether we
can remove part of the IPython completer, and more especially if we can
drop `python_matches`.

Once `python_matches` is dropped and some other of the current matchers
are either dropped or converted to the new API, that should simplify the
internal quite a bit.

That would just be too much for an already BIG pull-request.