Exit terminal interface gracefully on signals
Closes gh-9944

This sets signal handlers for SIGTERM and SIGHUP while the
prompt_toolkit prompt is active, letting it clean up the terminal before
the process exits. When the prompt is not active, we restore the
previous signals.

This can't work (as far as I know) if other code has set signal handlers
at the C level. I don't think this is likely to come up often, but
there's a new config option catch_signals=False which users can set to
disable our manipulation of signal handlers.