Backport PR #10191: IOStream: Ignore missing attrs from `dir()`. #6386
In some cases calling `getattr()` on attributes found with
`dir()` will fail. This is seen in Issue #6386.

The attribute in that issue is `name`, which is not needed for
`IOStream` to work correctly. This fix will set the attribute
to `None` and move on.

Tests were added in `IPython.utils.tests.test_io.py` to catch
any regressions.